refactor(shared): detectCycles を shared パッケージに移動 (#165)
UI パッケージの dependency-graph.ts から detectCycles と
buildDependencyEdges を shared に移動し、UI 側は re-export に変更。
CLI の doctor コマンドから shared 経由で利用可能にする。
