+++ packages/shared/src/index.ts
@@ -3,3 +3,4 @@ export * from "./schema.js";
 export * from "./constants.js";
 export * from "./comments.js";
 export * from "./status-dates.js";
+export * from "./dependency-graph.js";
